requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
requests
google-generativeai